lifting ddp override to ModelPT
