+++ nemo/core/classes/modelPT.py
@@ -24,10 +24,12 @@ from os import path
 from typing import Callable, Dict, List, Optional, Union
 
 import hydra
+from pytorch_lightning.overrides.data_parallel import LightningDistributedDataParallel
 import torch
 from omegaconf import DictConfig, OmegaConf, open_dict
 from pytorch_lightning import LightningModule, Trainer
 from pytorch_lightning.utilities import rank_zero_only
+from torch.nn.parallel.distributed import DistributedDataParallel
 
 from nemo.core import optim
 from nemo.core.classes.common import Model
@@ -1227,6 +1229,16 @@ class ModelPT(LightningModule, Model):
                 self.cfg = self._cfg
             else:
                 raise ValueError("`dataset_name` when updating config must be one of [train, validation, test]")
+
+    def configure_ddp(self, model: LightningModule, device_ids: List[int]) -> DistributedDataParallel:
+        logging.info('overriding ddp to set find_unused_parameters to True')
+        model = LightningDistributedDataParallel(model, device_ids=device_ids, find_unused_parameters=True)
+        return model
+
+    def setup(self, stage):
+        if stage == "fit":
+            # Update PTL trainer to use our configure_ddp
+            self._trainer.accelerator_backend.ddp_plugin.configure_ddp = self.configure_ddp
 
     @property
     def num_weights(self):
